ONNX + TensorRT

+++ src/main/java/org/reinforce4j/games/Connect4Service.java
@@ -6,7 +6,7 @@ import org.reinforce4j.core.Player;
 public enum Connect4Service implements GameService<Connect4> {
   INSTANCE;
 
-  private static final Connect4 INITIAL_STATE = new Connect4(new int[6][7], Player.ONE);
+  private static final Connect4 INITIAL_STATE = new Connect4(new byte[6][7], Player.ONE);
 
   private final int NUM_MOVES = 7;
   private final int NUM_FEATURES = 42;
@@ -23,7 +23,7 @@ public enum Connect4Service implements GameService<Connect4> {
 
   @Override
   public Connect4 newInitialState() {
-    return new Connect4(new int[6][7], Player.ONE);
+    return new Connect4(new byte[6][7], Player.ONE);
   }
 
   @Override
