+++ src/main/java/org/reinforce4j/games/TicTacToeService.java
@@ -6,13 +6,13 @@ import org.reinforce4j.core.Player;
 public enum TicTacToeService implements GameService<TicTacToe> {
   INSTANCE;
 
-  private static final TicTacToe INITIAL_STATE = new TicTacToe(new int[3][3], Player.ONE);
+  private static final TicTacToe INITIAL_STATE = new TicTacToe(new byte[3][3], Player.ONE);
   private final int NUM_MOVES = 9;
   private final int NUM_FEATURES = 9;
 
   @Override
   public TicTacToe newInitialState() {
-    return new TicTacToe(new int[3][3], Player.ONE);
+    return new TicTacToe(new byte[3][3], Player.ONE);
   }
 
   @Override
